% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Create a Histogram}
\usage{
histogram(data, var, bins = 10, col = "lightpink")
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{var}{The name of the numeric variable to create a histogram for.}

\item{bins}{The number of bins for the histogram (default is 10).}

\item{col}{The color of the histogram bars (default is "lightpink").}
}
\description{
This function generates a histogram for a numeric variable in a given data frame.
}
\examples{
data <- data.frame(x = rnorm(100))
hist1 <- histogram(data, var = "x", bins = 20, col = "green")
print(hist1)


}
